+++ picme-java/src/main/java/com/example/picmejava/model/dto/AtualizarAlbumDTO.java
@@ -9,6 +9,6 @@ import lombok.Setter;
 public class AtualizarAlbumDTO {
 
     private String titulo;
-    private Integer idTema;
+    private Long idTema;
 
 }
