+++ picme-java/src/main/java/com/example/picmejava/model/dto/CadastroAlbumDTO.java
@@ -9,6 +9,6 @@ public class CadastroAlbumDTO {
 
     private String titulo;
     private String descricao;
-    private Integer idTema;
-    private Integer idFotografo;
+    private Long idTema;
+    private Long idFotografo;
 }
